requirements.txt
apache-airflow==2.10.2
apache-airflow-providers-apache-spark
pyspark==3.5.0
minio
pandas
requests
psycopg2-binary
sqlalchemy
s3fs